requirements.txt
Django==2.1.3
django-debug-toolbar==1.10.1
mysqlclient==1.3.13
pkg-resources==0.0.0
pytz==2018.7
SQLAlchemy==1.2.14
sqlparse==0.2.4
uWSGI==2.0.17.1
